MAS-226 rsrScore is a double

+++ src/main/java/uk/gov/justice/digital/delius/data/api/OffenderAssessments.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class OffenderAssessments {
 
     @ApiModelProperty(value = "Risk of Serious Recidivism")
-    private Integer rsrScore;
+    private Double rsrScore;
     @ApiModelProperty(value = "Offender Group Reconviction Scale")
     private Integer OGRSScore;
